tedious tuning

+++ include/MCMC.hpp
@@ -4,7 +4,7 @@
 #include <vector>
 
 void Wang_Landau(int, std::vector<double>&);
-double multicanonical_sampling(int, const std::vector<double>&);
+long double multicanonical_sampling(int, const std::vector<double>&);
 double softmax(const std::vector<double>&, int);
 
 #endif
